排班测试问题

+++ biz/src/main/java/com/ldz/biz/module/service/impl/PbServiceImpl.java
@@ -1,26 +1,12 @@
 package com.ldz.biz.module.service.impl;
 
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Date;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import org.apache.commons.collections4.CollectionUtils;
-import org.apache.commons.lang.StringUtils;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import com.ldz.biz.module.bean.PbClXlmodel;
-import com.ldz.biz.module.bean.PbInfo;
-import com.ldz.biz.module.bean.XbXlPb;
-import com.ldz.biz.module.bean.clpbInfo;
+import com.ldz.biz.module.bean.*;
 import com.ldz.biz.module.mapper.ClClMapper;
 import com.ldz.biz.module.mapper.ClPbMapper;
 import com.ldz.biz.module.mapper.PbInfoMapper;
 import com.ldz.biz.module.model.ClCl;
 import com.ldz.biz.module.model.ClPb;
+import com.ldz.biz.module.service.ClService;
 import com.ldz.biz.module.service.PbService;
 import com.ldz.sys.base.BaseServiceImpl;
 import com.ldz.sys.exception.RuntimeCheck;
@@ -28,8 +14,20 @@ import com.ldz.sys.model.SysJg;
 import com.ldz.sys.model.SysYh;
 import com.ldz.sys.service.JgService;
 import com.ldz.util.bean.ApiResponse;
-
+import com.ldz.util.commonUtil.DateUtils;
+import org.apache.commons.collections4.CollectionUtils;
+import org.apache.commons.lang.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import tk.mybatis.mapper.common.Mapper;
+
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Date;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class PbServiceImpl extends BaseServiceImpl<ClPb, String> implements PbService {
@@ -37,6 +35,8 @@ public class PbServiceImpl extends BaseServiceImpl<ClPb, String> implements PbSe
 	private ClPbMapper entityMapper;
 	@Autowired
 	private JgService jgService;
+	@Autowired
+	private ClService clService;
 	@Autowired
 	private PbInfoMapper pbinfomapper;
 	@Autowired
@@ -54,9 +54,31 @@ public class PbServiceImpl extends BaseServiceImpl<ClPb, String> implements PbSe
 
 	@Override
 	public ApiResponse<String> saveEntity(ClPb entity) {
+		RuntimeCheck.ifNull(entity,"当前选择的排班车辆有误，请核实！");
+		RuntimeCheck.ifBlank(entity.getDate(),"排线时间不能为空");
+		RuntimeCheck.ifBlank(entity.getClId(),"车辆ID不能为空");
+		RuntimeCheck.ifBlank(entity.getXlId(),"线路ID不能为空");
+		Date pbDate=null;
+		try {
+			pbDate=DateUtils.getDate(entity.getDate(),"yyyy-MM-dd");
+		} catch (ParseException e) {}
+		if(pbDate==null){
+			RuntimeCheck.ifFalse(false,"排班时间格式异常");
+		}
+
 		clpbInfo clpbInfo = new clpbInfo();
-		clpbInfo.setDate(entity.getDate());
+		clpbInfo.setDate(pbDate);
 		clpbInfo.setClid(entity.getClId());
+
+		ClCl clCl=clService.findByOrgCode(entity.getClId());
+		RuntimeCheck.ifNull(clCl,"车辆信息有误，请核实！");
+		String sjId=clCl.getSjId();
+		RuntimeCheck.ifBlank(sjId,"该车辆未绑定司机，无法进行排班");
+		String clZt=clCl.getZt();
+		if(!StringUtils.equals(clZt,"00")){
+			RuntimeCheck.ifBlank(sjId,"该车辆状态异常，无法进行排班");
+		}
+
 
 		// 通过车辆id找到当天是否有排班线路信息
 		List<PbInfo> findXlCl = pbinfomapper.findXlCl(clpbInfo);
@@ -68,6 +90,10 @@ public class PbServiceImpl extends BaseServiceImpl<ClPb, String> implements PbSe
 			entity.setJgdm(user.getJgdm());
 			entity.setJgmc(org.getJgmc());
 			entity.setCjsj(new Date());
+			entity.setCph(clCl.getCph());//车牌号码
+			entity.setPbsj(pbDate);//排班时间
+			entity.setSj(clCl.getSjId());//司机ID
+			entity.setSjxm(clCl.getSjxm());//司机姓名
 			save(entity);
 			return ApiResponse.saveSuccess();
 		}
@@ -82,11 +108,15 @@ public class PbServiceImpl extends BaseServiceImpl<ClPb, String> implements PbSe
 			entity.setJgdm(user.getJgdm());
 			entity.setJgmc(org.getJgmc());
 			entity.setCjsj(new Date());
+			entity.setCph(clCl.getCph());//车牌号码
+			entity.setPbsj(pbDate);//排班时间
+			entity.setSj(clCl.getSjId());//司机ID
+			entity.setSjxm(clCl.getSjxm());//司机姓名
 			save(entity);
 			return ApiResponse.saveSuccess();
+		}else{
+			return ApiResponse.fail("车辆与线路已经关联，需求重启关联");
 		}
-
-		return ApiResponse.fail();
 	}
 
 	@Override
@@ -144,16 +174,32 @@ public class PbServiceImpl extends BaseServiceImpl<ClPb, String> implements PbSe
 
 	@Override
 	public ApiResponse<String> deleteByXlAndCl(String xlId, String clId,String date) {
+		RuntimeCheck.ifBlank(xlId,"线路ID不能为空");
+		RuntimeCheck.ifBlank(clId,"车辆ID不能为空");
+		RuntimeCheck.ifBlank(date,"排班时间不能为空");
 
+		Date pbDate=null;
+		try {
+			pbDate=DateUtils.getDate(date,"yyyy-MM-dd");
+		} catch (ParseException e) {}
+		if(pbDate==null){
+			RuntimeCheck.ifFalse(false,"排班时间格式异常");
+		}
+
+		int i=0;
 		clpbInfo clpbInfo = new clpbInfo();
-		clpbInfo.setDate(date);
+		clpbInfo.setDate(pbDate);
 		clpbInfo.setClid(clId);
 		clpbInfo.setXlid(xlId);
 		List<PbInfo> findXlCl = pbinfomapper.findXlCl(clpbInfo);
-
-		entityMapper.deleteByPrimaryKey(findXlCl.get(0).getId());
-
-		return ApiResponse.success();
+		if(findXlCl!=null&&findXlCl.size()==1&&findXlCl.get(0).getId()!=null){
+			i=entityMapper.deleteByPrimaryKey(findXlCl.get(0).getId());
+		}
+		if(i==0){
+			return ApiResponse.fail();
+		}else{
+			return ApiResponse.success();
+		}
 	}
 
 	public String getNowDate() {
@@ -163,7 +209,20 @@ public class PbServiceImpl extends BaseServiceImpl<ClPb, String> implements PbSe
 
 		return date;
 	}
-
+	@Override
+	public ApiResponse<List<ClClModel>> getAllNotPbClList(String xlId, String date){
+		RuntimeCheck.ifBlank(xlId,"线路ID不能为空");
+		RuntimeCheck.ifBlank(date,"排班时间不能为空");
+		Date pbDate=null;
+		try {
+			pbDate=DateUtils.getDate(date,"yyyy-MM-dd");
+		} catch (ParseException e) {}
+		if(pbDate==null){
+			RuntimeCheck.ifFalse(false,"排班时间格式异常");
+		}
+		List<ClClModel> clClList=clclmapper.getAllNotPbClList(xlId,pbDate);
+		return ApiResponse.success(clClList);
+	}
 	public static void main(String[] args) {
 		
 		SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
